fix: Reestrutura algumas rotas e métodos
