+++ src/routes.js
@@ -8,6 +8,8 @@ const {
     SignToTrack,
     GetUserTracks,
     DeleteUser,
+    GetTracks,
+    GetContentsToTrack,
 } = require("./controllers/users");
 const {
     AdminSignUp,
@@ -15,6 +17,7 @@ const {
     AdminAddTrack,
     AdminAddTrackContent,
     DeleteContent,
+    DeleteTrack,
 } = require("./controllers/admins");
 const {
     validateUserData,
@@ -23,11 +26,6 @@ const {
     validateEmailUser,
 } = require("./middlewares/users");
 const { validateAdminData } = require("./middlewares/admins");
-const {
-    GetTracks,
-    GetContentsToTrack,
-    DeleteTrack,
-} = require("./controllers/tracks");
 const { userSignUpSchema } = require("./schemas/userSignUpSchema");
 
 routes.post("/signup", validateUserData(userSignUpSchema), SignUp);
@@ -36,13 +34,13 @@ routes.put("/user/update", validateEmailUser, UpdateUser);
 routes.delete("/user/delete/:id", DeleteUser);
 routes.post("/user/sign_track/:track_id", SignToTrack);
 routes.get("/user/tracks/:id", GetUserTracks);
-routes.get("/tracks", GetTracks);
-routes.delete("/tracks/delete/:id", DeleteTrack);
-routes.get("/:id/contents", GetContentsToTrack);
+routes.get("/user/all_tracks", GetTracks);
+routes.get("/user/:id/contents", GetContentsToTrack);
 routes.post("/admin/signup", validateAdminData(userSignUpSchema), AdminSignUp); //DEVELOPMENT ONLY
 routes.post("/admin/login", AdminLogin);
 routes.post("/admin/add_track", AdminAddTrack);
 routes.post("/admin/add_content", AdminAddTrackContent);
+routes.delete("/admin/del_track/:id", DeleteTrack);
 routes.delete("/admin/del_content/:id", DeleteContent);
 
 module.exports = routes;
